requirements.txt
fastapi
uvicorn[standard]
pillow
ultralytics>=8.0.0
python-multipart
